instagram-reels-poster/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
SQLAlchemy==2.0.45
APScheduler==3.10.4
instagrapi==1.19.7
cryptography==42.0.5
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
Werkzeug==3.0.0
pydantic<2.0
Pillow>=8.1.1
moviepy>=1.0.3